1. If you spend too much time thinking about a thing, you'll never get it done

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -11,7 +11,7 @@
 
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int a;
+	size_t a;
 
 	for (a = 0; a < size; a++)
 		action(array[a]);
